уровни

+++ world/storage.lua
@@ -1,12 +1,13 @@
 local COMMON = require "libs.common"
 local CONSTANTS = require "libs.constants"
 local JSON = require "libs.json"
+local LEVELS = require "assets.levels.levels"
 
 local TAG = "Storage"
 
 local Storage = COMMON.class("Storage")
 
-Storage.VERSION = 3
+Storage.VERSION = 4
 Storage.AUTOSAVE = 30 --seconds
 Storage.CLEAR = CONSTANTS.IS_DEBUG and false --BE CAREFUL. Do not use in prod
 Storage.LOCAL = CONSTANTS.IS_DEBUG and true --BE CAREFUL. Do not use in prod
@@ -14,6 +15,7 @@ Storage.LOCAL = CONSTANTS.IS_DEBUG and true --BE CAREFUL. Do not use in prod
 function Storage:initialize()
     self:_load_storage()
     self.prev_save_time = os.clock()
+    self.save_on_update = false
 end
 
 function Storage:changed()
@@ -40,13 +42,13 @@ function Storage:_load_storage()
         else
             print("can't load from file:" .. tostring(read_err))
             self:_init_storage()
-            self:save()
+            self:save(true)
         end
     else
         self:_init_storage()
-        self:save()
     end
     self:_migration()
+    self:save(true)
     COMMON.i("loaded", TAG)
 end
 
@@ -55,10 +57,13 @@ function Storage:update(dt)
         COMMON.EVENT_BUS:event(COMMON.EVENTS.STORAGE_CHANGED)
         self.change_flag = false
     end
+    if (self.save_on_update) then
+        self:save(true)
+    end
     if (Storage.AUTOSAVE and Storage.AUTOSAVE ~= -1) then
         if (os.clock() - self.prev_save_time > Storage.AUTOSAVE) then
             COMMON.i("autosave", TAG)
-            self:save()
+            self:save(true)
         end
     end
 
@@ -70,28 +75,62 @@ function Storage:_init_storage()
         debug = {
             draw_level_matcher = false,
         },
+        level_current = 1;
+        levels = {
+
+        },
         version = Storage.VERSION
     }
+    self:_check_levels()
+
+end
+
+function Storage:_check_levels()
+    for i, level in ipairs(LEVELS.levels) do
+        local storage_level = self.data.levels[i]
+        if (storage_level == nil) then
+            storage_level = {
+                stars = 0
+            }
+            self:save()
+            self.data.levels[i] = storage_level
+        end
+    end
+    pprint(self.data)
 end
 
 function Storage:_migration()
     if (self.data.version < Storage.VERSION) then
         COMMON.i(string.format("migrate from:%s to %s", self.data.version, Storage.VERSION), TAG)
-        -- 1->2
+        -- 1->3
         if (self.data.version < 3) then
             self:_init_storage()
         end
 
+        -- 3->4
+        if(self.data.version < 4) then
+            self.data.levels = {}
+            self:_check_levels()
+        end
+
+
+        self:_check_levels()
         self.data.version = Storage.VERSION
     end
 end
 
-function Storage:save()
-    COMMON.i("save", TAG)
-    self.prev_save_time = os.clock()
-    local data = self:_get_path()
-    local file = io.open(data, "w+")
-    file:write(JSON.encode(self.data, true))
+function Storage:save(force)
+    if (force) then
+        COMMON.i("save", TAG)
+        self.prev_save_time = os.clock()
+        local data = self:_get_path()
+        local file = io.open(data, "w+")
+        file:write(JSON.encode(self.data, true))
+        self.save_on_update = false
+    else
+        self.save_on_update = true
+    end
+
 end
 
 return Storage
